пытается сделать презентер для карточки фильма
